+++ NEWS.md
@@ -1,6 +1,14 @@
 ucsf-vpn
 ========
 
+## Version 6.1.0 (2024-11-11)
+
+### Deprecated and Defunct
+
+ * Command-line option `--method=<value>` and corresponding
+   environment variable `UCSF_VPN_METHOD` are defunct.
+ 
+
 ## Version 6.1.0 (2024-06-26)
 
 ### Significant changes